python/pyproject.toml
[project]
name = "lca-production-monitoring"
version = "0.1.0"
description = "OfficeFlow agent with LangSmith production monitoring"
requires-python = "==3.12.0"
dependencies = [
    "langchain",
    "langchain-anthropic",
    "langchain-community",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "ddtrace",
    "openai>=2.29.0",
    "numpy>=2.4.2",
    "langsmith>=0.7.12",
    "click>=8.3.1",
    "rich>=14.3.3",
]

